chore: refactor cli (#90)
