+++ packages/alchemy/bin/commands/login.ts
@@ -1,0 +1,93 @@
+import * as Config from "effect/Config";
+import * as ConfigProvider from "effect/ConfigProvider";
+import * as Console from "effect/Console";
+import * as Effect from "effect/Effect";
+import * as Layer from "effect/Layer";
+import * as Logger from "effect/Logger";
+import { Command, Flag } from "effect/unstable/cli";
+
+import { AuthProviders } from "../../src/Auth/AuthProvider.ts";
+import {
+  getProfile,
+  setProfile,
+  withProfileOverride,
+} from "../../src/Auth/Profile.ts";
+import { Stage } from "../../src/Stage.ts";
+import * as State from "../../src/State/index.ts";
+import { loadConfigProvider } from "../../src/Util/ConfigProvider.ts";
+import { fileLogger } from "../../src/Util/FileLogger.ts";
+
+import { envFile, importStack, main, profile, stage } from "./_shared.ts";
+
+const loginConfigure = Flag.boolean("configure").pipe(
+  Flag.withDescription(
+    "Run the provider's interactive configure step before logging in",
+  ),
+  Flag.withDefault(false),
+);
+
+export const loginCommand = Command.make(
+  "login",
+  {
+    main,
+    envFile,
+    stage,
+    profile,
+    configure: loginConfigure,
+  },
+  Effect.fnUntraced(function* ({ main, stage, envFile, profile, configure }) {
+    const stackEffect = yield* importStack(main);
+
+    const authProviders: AuthProviders["Service"] = {};
+
+    const services = Layer.mergeAll(
+      Layer.succeed(AuthProviders, authProviders),
+      ConfigProvider.layer(
+        withProfileOverride(yield* loadConfigProvider(envFile), profile),
+      ),
+      Logger.layer([fileLogger("out")]),
+      Layer.succeed(Stage, stage),
+      State.LocalState,
+    );
+
+    yield* Effect.gen(function* () {
+      // Build the stack — this triggers each provider's AuthProviderLayer,
+      // which registers itself into the shared `authProviders` registry.
+      yield* stackEffect;
+
+      const ci = yield* Config.boolean("CI").pipe(Config.withDefault(false));
+      const providers = Object.values(authProviders);
+
+      if (providers.length === 0) {
+        yield* Console.log(
+          "No AuthProviders registered. Make sure the stack's providers() layer includes AuthProviderLayer entries.",
+        );
+        return;
+      }
+
+      yield* Effect.forEach(
+        providers,
+        (provider) =>
+          Effect.gen(function* () {
+            const existing = yield* getProfile(profile);
+            const stored = existing?.[provider.name];
+
+            let cfg: { method: string };
+            if (configure || stored == null) {
+              cfg = yield* provider.configure(profile, { ci });
+              yield* setProfile(profile, {
+                ...existing,
+                [provider.name]: cfg,
+              });
+            } else {
+              cfg = stored;
+            }
+
+            yield* provider.login(profile, cfg);
+            yield* provider.prettyPrint(profile, cfg);
+          }),
+        { discard: true },
+      );
+    }).pipe(Effect.provide(services));
+  }),
+);
